changed stayRoused to 16 bit fn for finer control

+++ greenlogger/greenlogger/src/ExtInterrupt/ExtInterrupts.h
@@ -9,7 +9,7 @@
 #ifndef EXTINTERRUPTS_H_
 #define EXTINTERRUPTS_H_
 
-void stayRoused(uint8_t sec);
+void stayRoused(uint16_t sec);
 void endRouse(void);
 void keepBluetoothPowered(uint8_t sec);
 void shutDownBluetooth(void);
